refactor: Move checking env to separate module
