+++ src/api/commercetools-client.ts
@@ -1,5 +1,3 @@
-'use strict';
-
 import {
   Client,
   ClientBuilder,
@@ -12,38 +10,10 @@ import {
   createApiBuilderFromCtpClient,
 } from '@commercetools/platform-sdk';
 
-const requiredEnvVars: string[] = [
-  'VITE_CTP_PROJECT_KEY',
-  'VITE_CTP_SCOPES',
-  'VITE_CTP_API_URL',
-  'VITE_CTP_AUTH_URL',
-  'VITE_CTP_CLIENT_ID',
-  'VITE_CTP_CLIENT_SECRET',
-];
-
-for (const envVar of requiredEnvVars) {
-  if (!import.meta.env[envVar]) {
-    throw new Error(`Missing environment variable: ${envVar}`);
-  }
-}
-
-function requireEnvVar<K extends keyof ImportMetaEnv>(name: K): string {
-  const value = import.meta.env[name];
-  if (!value) {
-    throw new Error(`Missing required environment variable: ${name}`);
-  }
-  return value;
-}
-
-export const projectKey = requireEnvVar('VITE_CTP_PROJECT_KEY');
-export const clientId = requireEnvVar('VITE_CTP_CLIENT_ID');
-export const clientSecret = requireEnvVar('VITE_CTP_CLIENT_SECRET');
-export const authUrl = requireEnvVar('VITE_CTP_AUTH_URL');
-export const apiUrl = requireEnvVar('VITE_CTP_API_URL');
-export const scopes = requireEnvVar('VITE_CTP_SCOPES');
+import env from './env';
 
 const httpMiddlewareOptions: HttpMiddlewareOptions = {
-  host: apiUrl,
+  host: env.apiUrl,
   includeRequestInErrorResponse: true,
   includeOriginalRequest: true,
   retryConfig: {
@@ -55,18 +25,18 @@ const httpMiddlewareOptions: HttpMiddlewareOptions = {
 
 const getClient = (): Client => {
   const authMiddlewareOptions: AuthMiddlewareOptions = {
-    host: authUrl,
-    projectKey: projectKey,
+    host: env.authUrl,
+    projectKey: env.projectKey,
     credentials: {
-      clientId: clientId,
-      clientSecret: clientSecret,
+      clientId: env.clientId,
+      clientSecret: env.clientSecret,
     },
-    scopes: scopes.split(','),
+    scopes: env.scopes.split(','),
     fetch: fetch,
   };
   return (
     new ClientBuilder()
-      .withProjectKey(projectKey)
+      .withProjectKey(env.projectKey)
       .withClientCredentialsFlow(authMiddlewareOptions)
       .withHttpMiddleware(httpMiddlewareOptions)
       .withUserAgentMiddleware()
@@ -77,4 +47,4 @@ const getClient = (): Client => {
 
 export const apiRoot: ByProjectKeyRequestBuilder = createApiBuilderFromCtpClient(
   getClient()
-).withProjectKey({ projectKey });
+).withProjectKey({ projectKey: env.projectKey });
